fix(graphql-api): silence most Kotlin compiler warnings
Opt graphql-api into param-property annotation targets, as server-core
already does, which clears the 791 warnings on DGS-generated
@JsonProperty constructor properties. Hand-written @Value/@Lazy
constructor parameters use an explicit @param: target to keep today's
injection behaviour.

Also drop needless !!, ?. and ?: on non-null receivers, redundant
nullable markers and conversions, and replace the deprecated
Locale(String) and StringUtils.startsWith calls.

+++ packages/http-api/src/main/kotlin/org/migor/feedless/http/mapper/HttpScrapeFlowMapper.kt
@@ -53,7 +53,7 @@ class HttpScrapeFlowMapper : StoredFlowParser {
       id = body.id?.let { SourceId(it) } ?: SourceId(),
       title = body.title,
       tags = body.tags?.toTypedArray(),
-      latLon = body.latLng?.let { LatLonPoint(it.lat.toDouble(), it.lng.toDouble()) },
+      latLon = body.latLng?.let { LatLonPoint(it.lat, it.lng) },
       actions = toDomainActions(body.flow),
       lastErrorMessage = body.lastErrorMessage,
     )
